+++ src/app/views/requested-loan/requested-loan.component.ts
@@ -5,6 +5,7 @@ import { Loan } from './../../models/loan.model';
 // App Services
 import { ContractsService } from './../../services/contracts.service';
 import { AvailableLoansService } from '../../services/available-loans.service';
+import { FilterLoansService } from '../../services/filter-loans.service';
 
 @Component({
   selector: 'app-requested-loan',
@@ -12,17 +13,36 @@ import { AvailableLoansService } from '../../services/available-loans.service';
   styleUrls: ['./requested-loan.component.scss']
 })
 export class RequestedLoanComponent implements OnInit {
+  winHeight: number = window.innerHeight;
   loading: boolean;
   available: any;
   loans = [];
   availableLoans = true;
   pendingLend = [];
+  filters = {
+    currency: undefined,
+    amountStart: null,
+    amountEnd: null,
+    interest: null,
+    duration: null
+  };
+  filtersOpen = undefined;
 
   constructor(
     private contractsService: ContractsService,
     private spinner: NgxSpinnerService,
-    private availableLoansService: AvailableLoansService
-  ) {}
+    private availableLoansService: AvailableLoansService,
+    private filterLoansService: FilterLoansService
+  ) { }
+
+  openFilters() {
+    this.filtersOpen = !this.filtersOpen;
+  }
+
+  onFiltered() {
+    this.spinner.show();
+    this.loadLoans();
+  }
 
   // Available Loans service
   upgradeAvaiblable() {
@@ -31,11 +51,16 @@ export class RequestedLoanComponent implements OnInit {
 
   loadLoans() {
     this.contractsService.getOpenLoans().then((result: Loan[]) => {
-      this.loans = result;
+
+      const filterLoans = this.filterLoansService.filterLoans(result, this.filters);
+      this.loans = filterLoans;
+
       this.upgradeAvaiblable();
       this.spinner.hide();
       if (this.loans.length === 0) {
         this.availableLoans = false;
+      } else {
+        this.availableLoans = true;
       }
     });
   }