+++ src/app/views/requested-loan/requested-loan.module.ts
@@ -1,11 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 // App Modules
 import { NgxSpinnerModule, NgxSpinnerService } from 'ngx-spinner';
 import { SharedModule } from '../../shared/shared.module';
 import { MaterialModule } from './../../material/material.module';
 // App Services
 import { ContractsService } from './../../services/contracts.service';
+import { FilterLoansService } from '../../services/filter-loans.service';
 // App Component
 import { RequestedLoanComponent } from './requested-loan.component';
 
@@ -14,14 +16,17 @@ import { RequestedLoanComponent } from './requested-loan.component';
     CommonModule,
     NgxSpinnerModule,
     MaterialModule,
-    SharedModule
+    SharedModule,
+    ReactiveFormsModule,
+    FormsModule
   ],
   declarations: [
     RequestedLoanComponent
   ],
   providers: [
     NgxSpinnerService,
-    ContractsService
+    ContractsService,
+    FilterLoansService
   ],
   exports: [
     RequestedLoanComponent